Motion requests into FIFO, passed to pathing driver events finish
An event loaded with a movement can be published from another task.
These events are caught by the motion task, the movement put into a FIFO queue.
Movements are pulled from the queue, sent to the path interpolator.
Once done, the next item in the queue should be processed.

Needs cleanup with regards to queue storage, sizing, and use of pointers in the movement -> event -> queue items.